Source code changes for 1.5.6

+++ FNPlugin/GameConstants.cs
@@ -76,7 +76,7 @@ namespace FNPlugin
 
         public const double ELECTRON_CHARGE = 1.602176565e-19;
         public const double ATOMIC_MASS_UNIT =  1.660538921e-27;
-        public const double STANDARD_GRAVITY = 9.80665;
+        public const float STANDARD_GRAVITY = 9.80665f;
 
         public const double dilution_factor = 15000.0;
         public const double LfoFuelThrustModifier = 2.2222;
